# docs/swagger.yaml
definitions:
...
  models.UsersDTO:
    properties:
      avatar:
        type: string
      description:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastConnectionDate:
        type: string
      lastName:
        type: string
      mail:
        type: string
      nickname:
        type: string
      password:
        type: string
      phoneNumber:
        type: string
      registerDate:
        type: string
      type:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/user/login:
    post:
      description: Se connecte à un utilisateur
      parameters:
      - description: Mail de l'utilisateur
        in: body
        name: mail
        required: true
        schema:
          type: string
      - description: Mot de passe de l'utilisateur
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: Retourne un token de connexion
          schema:
            type: string
        "400":
          description: Requête incorrecte - données invalides
          schema:
            type: string
        "409":
          description: Conflit - L'email ou le mot de passe existe déjà
          schema:
            type: string
      summary: User
      tags:
      - Connexion
  /api/user/register:
    post:
      description: Crée un nouvel utilisateur
      parameters:
      - description: User to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UsersDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Retourne l'utilisateur crée
          schema:
            $ref: '#/definitions/models.UsersDTO'
        "400":
          description: Requête incorrecte - données invalides
          schema:
            type: string
        "409":
          description: Conflit - L'email ou le mot de passe existe déjà
          schema:
            type: string
      summary: User
      tags:
      - Création
swagger: "2.0"